+++ sentry-kotlin-multiplatform/src/androidMain/kotlin/io/sentry/kotlin/multiplatform/SentryAndroid.kt
@@ -1,7 +1,23 @@
 package io.sentry.kotlin.multiplatform
 
+import android.content.Context
 import io.sentry.Sentry
+import io.sentry.android.core.SentryAndroid
+import io.sentry.kotlin.multiplatform.extensions.toAndroidSentryOptions
 import io.sentry.kotlin.multiplatform.protocol.SentryId
+import io.sentry.kotlin.multiplatform.Sentry as SentryKmp
+
+/**
+ * Sentry initialization with a context and option configuration handler.
+ *
+ * @param context Application context.
+ * @param configuration Options configuration handler.
+ */
+fun SentryKmp.init(context: Context, configuration: (SentryOptions) -> Unit) {
+    val options = SentryOptions()
+    configuration.invoke(options)
+    SentryAndroid.init(context, options.toAndroidSentryOptions())
+}
 
 internal actual object SentryBridge {
