+++ sentry-kotlin-multiplatform/src/appleMain/kotlin/io/sentry/kotlin/multiplatform/SentryApple.kt
@@ -1,10 +1,24 @@
 package io.sentry.kotlin.multiplatform
 
 import cocoapods.Sentry.SentrySDK
+import io.sentry.kotlin.multiplatform.extensions.toCocoaSentryOptions
 import io.sentry.kotlin.multiplatform.nsexception.asNSException
+import io.sentry.kotlin.multiplatform.nsexception.setSentryUnhandledExceptionHook
 import io.sentry.kotlin.multiplatform.protocol.SentryId
 import platform.Foundation.NSError
 import platform.Foundation.NSException
+
+/**
+ * Sentry initialization with an option configuration handler.
+ *
+ * @param configuration Options configuration handler.
+ */
+fun Sentry.start(configuration: (SentryOptions) -> Unit) {
+    val options = SentryOptions()
+    configuration.invoke(options)
+    SentrySDK.startWithOptionsObject(options.toCocoaSentryOptions())
+    setSentryUnhandledExceptionHook()
+}
 
 internal actual object SentryBridge {
 
@@ -40,7 +54,7 @@ internal actual object SentryBridge {
         return { cocoaScope ->
             val cocoaScopeImpl = cocoaScope?.let {
                 SentryScopeCocoaImpl(it)
-            }  
+            }
             val scope = cocoaScopeImpl?.let { SentryScope(it) }
             scope?.let { scopeCallback.invoke(it) }
         }
