--- sentry-kotlin-multiplatform/src/commonMain/kotlin/io/sentry/kotlin/multiplatform/SentryOptionsConfiguration.kt
@@ -1,5 +1,0 @@
-package io.sentry.kotlin.multiplatform
-
-fun interface OptionsConfiguration<T: SentryOptions> {
-    fun configure(options: T)
-}
